gh-108297: Remove test_crashers (#108690)
The test was skipped in 2011 by
commit 6d95b1bd87c81f1b217c93ad65da2b074a2bfdc4.

Scripts in Lib/test/crashers/ do not crash on a reliable way. They
rely on undefined behaviors, like state of the stack memory, and so
may or may not crash. It is not worth it to make sure that they crash
in a continious integration, they should be run manually time to time
instead.